requirements.txt
bitstring
python-telegram-bot